refactor: Extract per-session mechanism eligibility into SaslMechanismEligibility
Moves the narrowing from "mechanisms this deployment supports" to "mechanisms this session may use" out of SASLAuthentication, along with the per-user SCRAM lookup and its session cache. No behaviour change, and no delegates: nothing outside SASLAuthentication and its tests called any of it.

Moves the affected tests to match, splitting them into SaslMechanismEligibilityTest and SaslMechanismEligibilityScramTest.
