+++ xmppserver/src/test/java/org/jivesoftware/openfire/net/SASLAuthenticationTest.java
@@ -227,96 +227,6 @@ public class SASLAuthenticationTest
     }
 
     /**
-     * Verifies that getAvailableMechanismsForSession returns mechanisms for a ClientSession.
-     */
-    @Test
-    public void shouldReturnAvailableMechanismsForClientSession()
-    {
-        // Setup test fixture.
-        final Connection connection = mock(Connection.class);
-        when(connection.isEncrypted()).thenReturn(false);
-
-        final StreamID streamID = new BasicStreamIDFactory().createStreamID();
-        final LocalClientSession session = new LocalClientSession(Fixtures.XMPP_DOMAIN, connection, streamID, Locale.ENGLISH);
-
-        // Execute system under test.
-        final Set<String> mechanisms = SASLAuthentication.getAvailableMechanismsForSession(session);
-
-        // Verify result.
-        assertTrue(mechanisms.contains("PLAIN"), "Expected PLAIN to be available for an unencrypted client session.");
-        assertFalse(mechanisms.contains("EXTERNAL"), "Expected EXTERNAL not to be available for an unencrypted client session without a trusted cert.");
-    }
-
-    /**
-     * Verifies that getAvailableMechanismsForSession returns mechanisms for an incoming server session.
-     */
-    @Test
-    public void shouldReturnAvailableMechanismsForIncomingServerSession()
-    {
-        // Setup test fixture.
-        final Connection connection = mock(Connection.class);
-        when(connection.isEncrypted()).thenReturn(false);
-
-        final StreamID streamID = new BasicStreamIDFactory().createStreamID();
-        final LocalIncomingServerSession session = new LocalIncomingServerSession(Fixtures.XMPP_DOMAIN, connection, streamID, "remote.example.org");
-
-        // Execute system under test.
-        final Set<String> mechanisms = SASLAuthentication.getAvailableMechanismsForSession(session);
-
-        // Verify result.
-        assertTrue(mechanisms.isEmpty(), "Expected no mechanisms to be available for an unencrypted server session without a trusted cert.");
-    }
-
-    /**
-     * Verifies that getAvailableMechanismsForSession does not advertise EXTERNAL for an incoming server session
-     * when EXTERNAL is disabled in the global SASL mechanisms configuration.
-     *
-     * Regression test for: Stream features advertise EXTERNAL even when disabled in sasl.mechs
-     */
-    @Test
-    public void shouldNotAdvertiseExternalForIncomingServerSessionWhenDisabledGlobally()
-    {
-        // Save the original enabled mechanisms to restore after the test.
-        final Set<String> originalMechanisms = new HashSet<>(SaslMechanismCatalog.getEnabledMechanisms());
-
-        try {
-            // Setup test fixture: Disable EXTERNAL in the global mechanisms configuration
-            SaslMechanismCatalog.setEnabledMechanisms(Collections.singletonList("PLAIN")); // Only PLAIN, no EXTERNAL
-
-            final Connection connection = mock(Connection.class);
-            when(connection.isEncrypted()).thenReturn(true);
-
-            final StreamID streamID = new BasicStreamIDFactory().createStreamID();
-            final LocalIncomingServerSession session = new LocalIncomingServerSession(Fixtures.XMPP_DOMAIN, connection, streamID, "remote.example.org");
-
-            // Execute system under test.
-            final Set<String> mechanisms = SASLAuthentication.getAvailableMechanismsForSession(session);
-
-            // Verify result.
-            assertFalse(mechanisms.contains("EXTERNAL"), "Expected EXTERNAL not to be advertised when disabled in global mechanisms configuration, even for encrypted sessions.");
-        } finally {
-            // Restore state to prevent affecting other unit tests.
-            SaslMechanismCatalog.setEnabledMechanisms(new ArrayList<>(originalMechanisms));
-        }
-    }
-
-    /**
-     * Verifies that getAvailableMechanismsForSession handles unknown session types gracefully.
-     */
-    @Test
-    public void shouldReturnEmptySetForUnknownSessionType()
-    {
-        // Setup test fixture.
-        final LocalSession unknownSession = mock(LocalSession.class);
-
-        // Execute system under test.
-        final Set<String> mechanisms = SASLAuthentication.getAvailableMechanismsForSession(unknownSession);
-
-        // Verify result.
-        assertTrue(mechanisms.isEmpty(), "Expected empty set for an unknown session type.");
-    }
-
-    /**
      * Verifies that authenticationSuccessful generates an anonymous auth token for a client with no username.
      * For SASL1, the success element has no authorization-identifier.
      */
@@ -715,7 +625,7 @@ public class SASLAuthenticationTest
         final LocalClientSession session = new LocalClientSession(Fixtures.XMPP_DOMAIN, connection, streamID, Locale.ENGLISH);
 
         // Execute system under test.
-        final Set<String> advertisableSASLMechanisms = SASLAuthentication.getAdvertisableSASLMechanisms(session);
+        final Set<String> advertisableSASLMechanisms = SaslMechanismEligibility.getAdvertisableSASLMechanisms(session);
         final Element result = SASLAuthentication.asSASLMechanismsElementForClientSessions(advertisableSASLMechanisms, false);
 
         // Verify result.
@@ -742,7 +652,7 @@ public class SASLAuthenticationTest
         final LocalClientSession session = new LocalClientSession(Fixtures.XMPP_DOMAIN, connection, streamID, Locale.ENGLISH);
 
         // Execute system under test.
-        final Set<String> advertisableSASLMechanisms = SASLAuthentication.getAdvertisableSASLMechanisms(session);
+        final Set<String> advertisableSASLMechanisms = SaslMechanismEligibility.getAdvertisableSASLMechanisms(session);
         final Element result = SASLAuthentication.asSASLMechanismsElementForClientSessions(advertisableSASLMechanisms, false);
 
         // Verify result.
@@ -768,7 +678,7 @@ public class SASLAuthenticationTest
         final LocalClientSession session = new LocalClientSession(Fixtures.XMPP_DOMAIN, connection, streamID, Locale.ENGLISH);
 
         // Execute system under test.
-        final Set<String> advertisableSASLMechanisms = SASLAuthentication.getAdvertisableSASLMechanisms(session);
+        final Set<String> advertisableSASLMechanisms = SaslMechanismEligibility.getAdvertisableSASLMechanisms(session);
         final Element result = SASLAuthentication.asSASLMechanismsElementForClientSessions(advertisableSASLMechanisms, true);
 
         // Verify result.
@@ -794,7 +704,7 @@ public class SASLAuthenticationTest
         final LocalClientSession session = new LocalClientSession(Fixtures.XMPP_DOMAIN, connection, streamID, Locale.ENGLISH);
 
         // Execute system under test.
-        final Set<String> advertisableSASLMechanisms = SASLAuthentication.getAdvertisableSASLMechanisms(session);
+        final Set<String> advertisableSASLMechanisms = SaslMechanismEligibility.getAdvertisableSASLMechanisms(session);
         final Element result = SASLAuthentication.asSASLMechanismsElementForClientSessions(advertisableSASLMechanisms, true);
 
         // Verify result.
@@ -819,7 +729,7 @@ public class SASLAuthenticationTest
         final LocalIncomingServerSession session = new LocalIncomingServerSession(Fixtures.XMPP_DOMAIN, connection, streamID, "remote.example.org");
 
         // Execute system under test.
-        final Set<String> advertisableSASLMechanisms = SASLAuthentication.getAdvertisableSASLMechanisms(session);
+        final Set<String> advertisableSASLMechanisms = SaslMechanismEligibility.getAdvertisableSASLMechanisms(session);
         final Element result = SASLAuthentication.asSASLMechanismsElementForServerSessions(advertisableSASLMechanisms, false);
 
         // Verify result.
@@ -845,7 +755,7 @@ public class SASLAuthenticationTest
         final LocalIncomingServerSession session = new LocalIncomingServerSession(Fixtures.XMPP_DOMAIN, connection, streamID, "remote.example.org");
 
         // Execute system under test.
-        final Set<String> advertisableSASLMechanisms = SASLAuthentication.getAdvertisableSASLMechanisms(session);
+        final Set<String> advertisableSASLMechanisms = SaslMechanismEligibility.getAdvertisableSASLMechanisms(session);
         final Element result = SASLAuthentication.asSASLMechanismsElementForServerSessions(advertisableSASLMechanisms, false);
 
         // Verify result.
@@ -870,7 +780,7 @@ public class SASLAuthenticationTest
         final LocalIncomingServerSession session = new LocalIncomingServerSession(Fixtures.XMPP_DOMAIN, connection, streamID, "remote.example.org");
 
         // Execute system under test.
-        final Set<String> advertisableSASLMechanisms = SASLAuthentication.getAdvertisableSASLMechanisms(session);
+        final Set<String> advertisableSASLMechanisms = SaslMechanismEligibility.getAdvertisableSASLMechanisms(session);
         final Element result = SASLAuthentication.asSASLMechanismsElementForServerSessions(advertisableSASLMechanisms, true);
 
         // Verify result.
@@ -895,7 +805,7 @@ public class SASLAuthenticationTest
         final LocalIncomingServerSession session = new LocalIncomingServerSession(Fixtures.XMPP_DOMAIN, connection, streamID, "remote.example.org");
 
         // Execute system under test.
-        final Set<String> advertisableSASLMechanisms = SASLAuthentication.getAdvertisableSASLMechanisms(session);
+        final Set<String> advertisableSASLMechanisms = SaslMechanismEligibility.getAdvertisableSASLMechanisms(session);
         final Element result = SASLAuthentication.asSASLMechanismsElementForServerSessions(advertisableSASLMechanisms, true);
 
         // Verify result.
@@ -1271,96 +1181,6 @@ public class SASLAuthenticationTest
     }
 
     /**
-     * Verifies that channel-binding types are not advertised when no channel-binding-capable mechanism is offered,
-     * since a peer that cannot use channel binding has no use for the type list.
-     *
-     * The connection is deliberately stubbed to support types: this asserts that the mechanism list alone is
-     * enough to suppress the announcement, rather than the outcome coinciding with a connection that happens to
-     * support nothing.
-     *
-     * @see <a href="https://xmpp.org/extensions/xep-0440.html">XEP-0440: SASL Channel-Binding Type Capability</a>
-     */
-    @Test
-    public void getAdvertisableChannelBindingTypes_noPlusMechanism_returnsEmpty()
-    {
-        // Setup test fixture.
-        final Connection connection = mock(Connection.class);
-        when(connection.getSupportedChannelBindingTypes()).thenReturn(Set.of("tls-exporter"));
-        final LocalSession session = mock(LocalSession.class);
-        when(session.getConnection()).thenReturn(connection);
-
-        // Execute system under test.
-        final Set<String> result = SASLAuthentication.getAdvertisableChannelBindingTypes(session, Set.of("PLAIN", "EXTERNAL"));
-
-        // Verify result.
-        assertTrue(result.isEmpty(), "Expected no channel-binding types when no -PLUS mechanism is offered.");
-    }
-
-    /**
-     * Verifies that no channel-binding types are advertised when a -PLUS mechanism is offered but the session's
-     * connection cannot supply any type in its current state (for example, because it is not encrypted).
-     */
-    @Test
-    public void getAdvertisableChannelBindingTypes_plusMechanismButConnectionSupportsNone_returnsEmpty()
-    {
-        // Setup test fixture.
-        final Connection connection = mock(Connection.class);
-        when(connection.getSupportedChannelBindingTypes()).thenReturn(Set.of());
-        final LocalSession session = mock(LocalSession.class);
-        when(session.getConnection()).thenReturn(connection);
-
-        // Execute system under test.
-        final Set<String> result = SASLAuthentication.getAdvertisableChannelBindingTypes(session, Set.of("SCRAM-SHA-1", "SCRAM-SHA-1-PLUS"));
-
-        // Verify result.
-        assertTrue(result.isEmpty(), "Expected no channel-binding types when the connection supports none.");
-    }
-
-    /**
-     * Verifies that the channel-binding types that the session's connection supports are advertised when a -PLUS
-     * mechanism is offered.
-     *
-     * The types come from the connection rather than from the globally registered providers, because
-     * Connection#getSupportedChannelBindingTypes is defined to reflect the connection's current state. Advertising
-     * a type the connection cannot supply would leave a peer that selects it to be rejected later, when the
-     * channel-binding data cannot be produced.
-     */
-    @Test
-    public void getAdvertisableChannelBindingTypes_plusMechanismWithConnectionSupport_returnsConnectionTypes()
-    {
-        // Setup test fixture.
-        final Connection connection = mock(Connection.class);
-        when(connection.getSupportedChannelBindingTypes()).thenReturn(Set.of("tls-server-end-point", "tls-exporter"));
-        final LocalSession session = mock(LocalSession.class);
-        when(session.getConnection()).thenReturn(connection);
-
-        // Execute system under test.
-        final Set<String> result = SASLAuthentication.getAdvertisableChannelBindingTypes(session, Set.of("SCRAM-SHA-1", "SCRAM-SHA-1-PLUS"));
-
-        // Verify result.
-        assertEquals(Set.of("tls-server-end-point", "tls-exporter"), result, "Expected the connection's supported channel-binding types to be advertised alongside a -PLUS mechanism.");
-    }
-
-    /**
-     * Verifies that no channel-binding types are advertised for a session that has no connection, rather than the
-     * lookup failing. A stream-management session that is detached has no connection, and stream features can be
-     * regenerated after it resumes.
-     */
-    @Test
-    public void getAdvertisableChannelBindingTypes_noConnection_returnsEmpty()
-    {
-        // Setup test fixture.
-        final LocalSession session = mock(LocalSession.class);
-        when(session.getConnection()).thenReturn(null);
-
-        // Execute system under test.
-        final Set<String> result = SASLAuthentication.getAdvertisableChannelBindingTypes(session, Set.of("SCRAM-SHA-1", "SCRAM-SHA-1-PLUS"));
-
-        // Verify result.
-        assertTrue(result.isEmpty(), "Expected no channel-binding types for a session that has no connection.");
-    }
-
-    /**
      * Verifies that exactly one XEP-0440 capability element is advertised, even when both the SASL1 and SASL2
      * feature elements are offered.
      *
@@ -1707,7 +1527,7 @@ public class SASLAuthenticationTest
         final LocalClientSession session = new LocalClientSession(Fixtures.XMPP_DOMAIN, connection, streamID, Locale.ENGLISH);
 
         // Execute system under test.
-        final Set<String> advertisableSASLMechanisms = SASLAuthentication.getAdvertisableSASLMechanisms(session);
+        final Set<String> advertisableSASLMechanisms = SaslMechanismEligibility.getAdvertisableSASLMechanisms(session);
         final Element result = SASLAuthentication.asSASLMechanismsElementForClientSessions(advertisableSASLMechanisms, true);
 
         // Verify result.
@@ -1739,7 +1559,7 @@ public class SASLAuthenticationTest
         final LocalClientSession session = new LocalClientSession(Fixtures.XMPP_DOMAIN, connection, streamID, Locale.ENGLISH);
 
         // Execute system under test.
-        final Set<String> advertisableSASLMechanisms = SASLAuthentication.getAdvertisableSASLMechanisms(session);
+        final Set<String> advertisableSASLMechanisms = SaslMechanismEligibility.getAdvertisableSASLMechanisms(session);
         final Element result = SASLAuthentication.asSASLMechanismsElementForClientSessions(advertisableSASLMechanisms, true);
 
         // Verify result: either no <inline/> or <inline/> without a <fast/> child.
